Drop unneeded partition dropping part of oozie import
As dropping partitions is handled differently in refinery, the scripts
for it are not needed. Also, the path names are still focused on the
precursor of refinery. Furthermore, SerDe handling has been
simplified.

Note that the part for adding partitions is currently not working
either. We keep it around nonetheless, as follow-up commits will bring
it into shape.

Bug: 67128
Change-Id: I60f5ce894a1a8d69a599b6f5b3a788c9859b555e