Adapt default auxpath for cluster setup
Hive currently grabs the auxpath from the site wide configuration, so
we no longer need to pass it per default.
Also, the default path currently does not exist for our machines.
Hence, we remove the default to avoid having it get in our way.

But we leave the mechanism to apply options in place.

Change-Id: If17dccb09aad6c736dd8ffc99729553c04f6d938