Stop zero-padding partition specifications
At least with CDH 5, values in partition specifications get treated as
strings. When zero-padding partitions specifications (e.g.: using 05
for month May instead of 5), the partition has to be matched again
using strings with zero padding [1]. This is counter-intuitive, and
got in the way at least twice when prototyping.

Hence, we stop to zero-pad partition specifications, and as a result,
the matching becomes more natural [2].

[1] Like a partition added by month=05 can be matched by

* [...] WHERE month = '05'

but neither of

* [...] WHERE month = 5
* [...] WHERE month = '5'
* [...] WHERE month = 05

(note that the last item is the same as the specification, but still
does not match).

[2] Like a partition added by month=5 can be matched by any of

* [...] WHERE month = 5
* [...] WHERE month = '5'
* [...] WHERE month = 05

but

* [...] WHERE month = '05'

will not match it.

Change-Id: I28aa564011fcb148e71ba63fa94ec950ad8b0f80