Stop using potentially uninitialized variables in oozie helper workflows
The workflow to add partitions used the “webrequest_source” variable,
although this variable was not a parameter to the workflow. As the
variable is set in the currently only coordinator that runs this
workflow, it is typically available at run-time. But this setting
creates an invisible coupling between the (meant to be) isolated
workflow and the coordinator. To avoid this invisible coupling, we
remove the unneeded “webrequest_source” from the workflow name.

(This decision was discussed in IRC

  http://bots.wmflabs.org/~wm-bot/logs/%23wikimedia-analytics/20140721.txt

betwee [16:08:20] and [16:10:03])

Change-Id: I397d161838596f4fb9b62bb1c0b2ac0690c9dcf2