Stop tying Oozie's “data_directory” to webrequests table
Oozie's “data_directory” setting was pointing to the directory of a
webrequest_source partition. Both this naming and the tying in the
webrequest_source partition would have caused problems for the
upcoming addition of webstastcollector.

The naming was in the way, because when including the webrequest's
dataset definitions in webstatscollector, they would have required the
global “data_directory” variable to point to the /webrequests/ data
directory.

Including the webrequest_source partition in the “data_directory”
setting was nice when only dealing with a single source per
coordinator. But for webstatscollector, we need to pull in data from
both text and mobile sources in one go. This would not easily be
possible when having the webrequest_source into “data_directory”.

Hence, we construct the dataset's locations from a more general
data_directory, materialize the data set definitions for the
webrequest sources, and use parametrized names to refer to them.

Change-Id: I4f3b00089728627a8c36de4a9b184ef0bc0691c6