Tag property descriptions as 'description' in worklow files
Some descriptions in Oozie xml files have been tagged as 'value'
before, and hence have been treated as default values. This
mis-tagging did not cause errors, as we explicitly pass values for the
corresponding properties, and hence override those defaults. But we
now tag the descriptions as 'description' to reflect their intention.

Change-Id: I597fdd8ea474c801cd2adf618a3e4a9b2d6658a9