Abort deployment, if Oozie's Hive config seems to contain passwords
Analytics1027 is used for many maintenance tasks around
refinery. However, it should not be used to deploy refinery, as the
hive-site.xml on analytics1027 contains passwords.
To guard against refinery deployments from analytics1027, we check if
Oozie's hive config seems to contain passwords. If it does, we abort
the deployment. If it doesn't, we deploy.

Bug: T76806
Change-Id: I01f7f8b0a2195cba4083fab84be32f8de82f93af