Stop enforcing content file being 000000_0 for Oozie's archive workflow
With the upcoming legacy tsv generation, the content file will not
always be 000000_0, but could also be 000000_1, or 000000_2. In
testing, we saw the final number reaching 7. Hard-coding all possible
variants made Oozie's archive workflow even more spaghetti and more
ugly than it has already been.
To avoid this, we switch to identifying the content file using a shell
script that parses the output of 'hdfs dfs -ls'. That output is not
super-stable, but the code is order of magnitudes more readable than
the spaghetti alternative. And also, we're only relying on the
trailing part of the output lines, which makes filename detection a
bit less fragile.

Change-Id: I0c5a05f19e04a895db6ea2b77f3d6b50c6b8b7f7