Remove dependence on bits for Oozie's legacy TSVs jobs
Bits is about to temporary get turned off, so camus won't be
generating the directories for the webrequest_bits_unchecked
dataset. This would block all TSVs. To avoid such a blocking, and
since no TSV is using bits anyways at this point, we just remove the
dependence on bits for now.

Change-Id: I6ce9b0abb58728a529b5e171994772c72f91cc29