Put refined datsets definition in separate file
The addition of the variable $webrequest_refined_data_directory was causing
existing oozie properties to fail, since there is no default value
for this directory.

Once the refined dataset is considered stable, I'd like to refactor a little
bit of this.  Particularly, the 'refined' dataset should be the main one,
and the raw ones should be refered to as 'raw'.

Change-Id: I9337490755355a97873509ff53e222a0f0db80c2