Fix legacy_tsvs' use of datasets after switch to refined tables
The switching to refined tables as default data source broke the
dataset dependencies of the legacy_tsvs, as webrequest_*_unchecked
became webrequest_*_raw_unchecked.

Simply following this move would allow to run the legacy_tsv scripts
again, but would buy a dependency of both the raw and refined
datasets--which we do not want. Unconditionally depending on all
refined webrequest_sources would work, but would make turning off a
single webrequest_source in camus harder.

Hence, we split out the different kinds of dependencies into separate
coordinators. Altough this approach buys a bit of maintenance burden,
but the dataset dependencies are really to the point, which will help
when turning webrequest_sources on/off.

Change-Id: I666d7692e9713954e3069fbf0eb4975d4a6b92ea