Geocode glam_nara files
Geocoding happens directly on the client ip without resolving
X-Forwarded-For. While that's wrong from the start, it is what the
tsvs on the udp2log pipeline do and did. And as at this point the
objective is switching away from udp2log, fixing the conceptual issues
with the glam_nara tsvs is left for the future.

Change-Id: I54f5db3f61291b0d455e44fd20de090b40c1a3ef