Deduplicate legacy-tsvs webrequest_source dependency information
Which of the legacy-tsvs depended on which webrequest_source was
duplicated by having the dependencies in the used coordinator, and
also in the HiveQL files. This was both not too nice maintenance wise,
and also it hinders reuse of the upcoming splitting out of 5xx tsvs
per webrequest_source.

We now extract the needed webrequest_sources from the coordinator
name. And thereby ease reuse of HiveQL files for different
webrequest_sources.

But HiveQL files need not rely on this. They can still choose to
duplicate webrequest_source dependency information.

Change-Id: I5ea5b3c874b3f94763413535077126469a77a9b1