Add per webrequest_source 5xx tsvs to legacy_tsvs
Ops called out that they would not want to see results from the misc
cluster when grepping for production issues, and this probably was an
oversight to have them in the udp2log tsvs. But they'd also voiced
that generally, having 5xx for misc might be useful.

As filtering tsvs for their webrequest_sources is generally hard, we
split all of them out into separate files. That way, Ops can grep for
production issues without having to deal with misc. They for example
can easily focus on only text or only mobile. And they also have
prepared (separate) files for misc's 5xx.

Change-Id: Id2adfa9ddb60789210a36b5472878890de580ace