Improve the format of the browser report
Makes the browser report output to an external table instead of
a set of files. Also splits the data into some new columns and
changes the percentage to an absolute value of view counts.

In addition to that, adds the long tail of the report (that was
previously cut off) to the data by collapsing it into a single
row where all dimension values are 'unknown'. Thus preserving
privacy and ensuring that the data doesn't get too big.

Bug: T126282
Change-Id: I19074c3f449a81c0eeeeb666964bbcfae0f05363