Remove some old refinery-*.jar artifacts
Refinery deployments are taking a bit of space, and now that scap
is being used for deploys, multiple copies of it will hang around.

Since refinery is deployed to HDFS, and Oozie jobs are started against
specific versioned HDFS deployment paths, this should be relatively safe.
I don't see any refinery_jar_version in oozie/ different than 0.0.24
or 0.0.31.

There is one case where things could break.  If a long running oozie job
has refinery_jar_version set to a version of a jar that is being removed,
if that job is stopped and then restarted using a newer versioned HDFS
refinery path, the old refinery_jar_version jar won't exist, and the job
will fail.  This would be obvious when the job runs, so we should be able
to fix soon.

Change-Id: I29037bbab4fbdda37bd85d02e8d6163ab79edbcd