Update hive-site.xml references
hive-site.xml is needed on hdfs for oozie to run hive queries.
We used to copy it onto hdfs as part of the refinery-deploy script.
This approach was causing problems particularly when testing (user-
defined refinery folder have not been deployed using the deploy
script, therefore don't have the hive-site.xml file).
Now hive-site.xml is copied onto hdfs (hdfs:///user/hive/hive-site.xml)
via puppet, and this patch is changing oozie job definition to match
that path, and to remove copying the file in the deploy script.

Bug: T133208
Change-Id: I9e82874164c91602206ae87f5642c05627a6af28