Improve oozie data loss alarms
To avoid false alarms when webrequest records do not have a
timestamp, does not consider records with dt='-' when creating
sequence statistics.

To monitor that the number of records with dt='-' doesn't get
too big, adds the count of such records to the statistics table.

The extract_data_loss check now also looks at the dt='-' records
and also notifies the error/warning if the lost/incorrect data
percentage surpasses the given threshold.

Also, tries to improve the error messages.

And updates the creation statements for the statistics tables.
Probably, an alter table on those tables will be needed when
deploying.

Bug: T148980
Change-Id: I89b8df63a39347d62347f67ec563d2c81a086f07