Add banner activity oozie jobs
Adds a banner_activity folder to the oozie folder in refinery.
It contains a duid folder that contains 2 jobs: daily and monthly.
The daily job calculates banner activity for a day and loads
it into druid in an minutely resolution (1-day segments).
The monthly job calculates banner activity for a month and
loads it into druid also in an minutely resolution, and overwrites
the data generated by the daily jobs. The monthly job's segments
are 1-month big, giving druid a better compression factor.

Bug: T155141
Change-Id: Iae85a39048596c71a8b4838a1b4ee60b9f6f6b7f