Update sqoop script with labsdb specificity
Compared to production, wikis in labsdb are views instead of databases.
Their name reflect this change having a '_p' postfix.
This patch adds a --labsdb parameter to the sqoop script, adding the
needed postfix to database name when parameter is set and setting
empty strings as default values for ar_content_format and
ar_content_model fields of the archive table.

The patch also includes a new --processors parameter allowing
to set how concurrent sqoop jobs are launched (labsdb restricts
the number of parallel connection for a user to 10).

Bug: T155658
Change-Id: I91a4d0892bfe7ce5ef55035ffab3077845f79288