Adjust drop hourly script to handle mediawiki log partitions
Log data in the mediawiki -> kafka -> camus -> hdfs pipeline has
various PII in it that needs to be deleted after no more than 90
days. The existing refinery-drop-hourly-partitions needed only minor
adjustments to handle this new type of data.

There are no puppetized uses of the 'raw' partition type, and it almost
matches the way camus writes data out. As such rename raw to camus to
match the way data is written out, and how external table LOCATION keys
in hive reference the data. Best practice should still be to explicitly
name the location of data, but this allows auto-detection to work.

Change-Id: Ia6811a7447ff63c2b0b78ed4a90d801e700a9379