Add snapshot to sqoop and namespace_map scripts
Sqoop and namespace_map scripts will likely use 'snapshot' as
a way to version imported data. Convention is that 'snapshot'
will be YYYY-MM for regular labs import, and will have a postfix
for other imports.
This patch modifies the sqoop and namespace_map scripts so that
they can accept 'snapshot' parameter and create partition-folders
accordingly.

Bug: T160152
Change-Id: I4c03aee6a5964900f1388506f660273152d6993d