Update static wiki projects list
Provide two lists, for prod and labs, instead of one.
Projects available for sqooping are different in prod
and labs. Some projects in prod are not in labs
(labswiki and labstestwiki), and sometimes labs can be
late in replicating data to its db infrastructure
(currently the case). Those files allow for distinct
sqoop jobs.

Bug: T160153
Change-Id: I8f2591230dcba05c0775a67e72a306fbe4caa78b