Fix domain_abbrev_map job to disambiguate wikimedia projects
This script was generating ambiguous records for <lang>.wikimedia
projects. For example: dbname = arwikimedia, for which was
generating ar.m, which is the same as arwiki for mobile site.
Hence the resulting map table was not deterministic and wrong.
The documentation in
https://wikitech.wikimedia.org/wiki/Analytics/Data/Pagecounts-all-sites
makes it clear that the .m suffix for xyz.wikimedia sites is only
considered for the following wikis:
commons.wikimedia.org
meta.wikimedia.org
incubator.wikimedia.org
species.wikimedia.org
strategy.wikimedia.org
outreach.wikimedia.org
usability.wikimedia.org
quality.wikimedia.org
SO, this change removes all <lang>.wikimedia.org from the table
because they are not considered (and hence irrepresentable) by the
webstatscollector abbreviation system.

Bug: T156388
Change-Id: I041f5463de24d5c250c176af684ece11ed65b533