Use both projectcounts raw and all sites to load cassandra
This patch modifies the hql query that gathers the data to load
into cassandra for AQS's legacy pagecounts per project endpoint.
It uses now both projectcounts_raw and projectcounts_all_sites
tables to crunch the data. On the one hand, pagecounts_raw is the
data set that spans the longest (from 2007 to 2016), but it has
no mobile-site counts. On the other hand, pagecounts_all_sites
has mobile-site counts, but only covers 2014 to 2016.

Bug: T161494
Change-Id: Ic4b4c0c855a75cb89493e65beefb7d9355307aa0