Update sqoop script to better handle failures
Currently when one or more table fail to be loaded by the sqoop script,
the global job fails (as expected) but doesn't provide much
information. This patch adds a retry feature to the script,
adds HDFS paths checks and force features, and finally logs better
information in case of failure.

Change-Id: If40acfcdc1948f1feb80b66d283f8a957e1811f3