Add --generate-jar and --jar-file options
This change adds the option to generate a jar file containing all the
ORM java code that sqoop would otherwise have to generate by hitting the
source.  It also adds the option to use that generated file.  It also
adds the jar that was generated during testing to artifacts/

Open Question: sqoop is accepting the --jar-file parameter and uses it:

INFO tool.CodeGenTool: Using existing jar: ...jar

However, it's still executing a SQL query and generating the avsc file
for each table.  I wonder if this is just unavoidable.

Deployment plan:
* (DONE) generate the jar, commit and merge to this repo
* (DONE) test running the sqoop with a --jar-file
* change the sqoop cron in puppet to use the new jar

NOTE: to compile these ORM .java files manually, the command is:
javac -classpath \
 /usr/lib/hadoop/* \
:/usr/lib/hadoop-0.20-mapreduce/* \
:/usr/lib/hadoop-hdfs/* \
:/usr/lib/sqoop/* \
  <<pattern-for-all-your-table-ORM-files>>.java

Bug: T143119
Change-Id: I984ce686f65ac61802d07812a6f17dba9a92fe3e