Add unique devices project-wide oozie jobs
We now have a global last-access cookie [WMF-Last-Access-Global].
New jobs is this patch compute daily and monthly unique devices
project-wide and stores a less detailed (without country)
version of it in the archive folder for external dumps.
This patch also includes a druid indexation job for daily
unique devices project-wide.

Bug: T143928
Change-Id: I60c9a068d6e962c3afa8519739d976c529aa920b