Update per host last access uniques oozie jobs
This patch reconciliates original per-host-uniques with the
new global uniques in term of naming and archiving format.
More precsiely, the old per-hosts job was archiving only the
unique_estimate value, while we found it actually makes sense
to provide also uniques_underestimate and uniques_offset (as it
is already the case in the new global uniques jobs).

Bug: T164597
Change-Id: If681a2c37f78e562e7db4fbf1ab5453c79c3dcc5