Update mediawiki history oozie job SLA
We recently changed the jobs to behave the same way as our usual
traffic jobs: coordinator waits for the full period for data to be
present instead of computing on data already loaded.
This induces a change in SLA, since now we accept to wait one month
plus 5 days for the monthly job instead of just 5 days.

Bug: T164713
Change-Id: I7b3fd20d809b5846685dd2f1ff892f111bff07cd