Create Oozie job for interlanguage nav table
We are aggregating pageviews from webrequest that are referred from one
language to another inside the same project family.  For example, if
someone navigates from en.wikipedia to de.wikipedia, we count that as
(wikipedia, en, de, 1).  The language team has been doing this manually
so far and we are just oozifying it so it can be monitored, reliable,
and automated.

Bug: T170764
Change-Id: I35bc02e046931735f89c0e7bcda676da172a81aa