Enable newly available wikis for sqooping
According to https://phabricator.wikimedia.org/T165233#3486500 we now
have all but 12 (and the labs test) wikis available to sqoop in the labs
cluster.  This enables all currently available wikis for the sqoop cron
to pick them up.  This cron runs on the 2nd of every month, so we're
merging this now and it will take effect on August 2nd.

Bug: T165233
Change-Id: Ib215c7452f7bcafee809fb9a98af5a289afd02b2