Add clickstream oozie job
This new oozie job runs monthly and by default stores its
resulting datasets in
/wmf/data/archive/clickstream/YYYY-MM/clikstream-WIKI-YYYY-MM.tsv.gz
for the computed month and wikis.
It uses the util/spark/submit subworkflow to launch the
scala application computing the datasets with a hive-context, and
the new oozie/util/loop workflow to loop over project list
to archive datasets.

Bug: T175844
Change-Id: Ie3466e19f51f59ba3d2ea6176b9b52bf35cb34b4