Add refinery-drop-hive-partitions
This script works with multiple tables in a database, and
can infer all partitions, not just hourly date time bucketed partiions,
to drop.

This adds a refinery.util.HivePartiion class to make
working with Hive partitions a little easier.  More work
could be done to make HiveUtils better with this.

NOTE: This is pretty inefficient to run on lots of tables,
as each hive CLI command is run via the shell, which requires
starting up a JVM each time we run a hive command.

Bug: T181064

Change-Id: I935551c63f35f747397aa05783f767925ab923bc