Point data and success flags to the same directory
While working on re-starting this job, I noticed the output data was
actually going to a different directory than the success flags.  I
missed this in manual oozie testing because it has to do with the table
definition, and I was using a different test table in my own db.  I'm
not sure how to test this in production, so a sanity check from someone
doing the code review would be nice.

Bug: T170764
Change-Id: I54e5d2017cd6baeb5ec2c1ba0764107b97b3afeb