Correct the column order
The interlanguage_navigation table had current and previous columns in
an order different from the insert/select statement that updates the
table.  This should have worked because the columns in the select are
labeled but apparently hive doesn't work like that, so I corrected the
data manually and this commit makes the table agree with the insert
statement (it's easier this way, because it doesn't require re-starting
or re-running any jobs).

Change-Id: Ida44ca3bab93d73fd10cdb97b776fdc047807da2