Add optional datasource to druid loading workflow
Druid loading workflow and python script were not allowing
to specify the datasource to load programmatically (it was
to be hard-coded in the loading template).
This patch adds the datasource as an optional parameter to
the loading script.
WARNING: If no value is provided but the template expects
one, datasource will be set to empty.

Change-Id: I40049532d2e9f1726fe0d4d9c213d93f1ff19f30