Refactor sqoop jobs and add from_timestamp
* split up code into a common module and different sqoop jobs
* add cu_change query
* test sqoop-generate-jar
* test sqoop-mediawiki-tables
* test sqoop-mediawiki-tables with cu_changes table
* factor out and improve logging config

Note: the logging_setup module now has configure_logging which takes a
logger, level, and optional log_file.  This sets up WARN and above to go
to the console (sys.stderr) and everything to go to the log_file if
specified.  Hopefully this simplifies logging and error reporting for
cron job setup.

Bug: T184759
Change-Id: Ibd8d9c603243c1063b154f1d54d9a618edca7190