Repair successfuly sqooped cu_changes
Once sqooped, raw tables coming from mediawiki have a _SUCCESS flag.
They need a _PARTITIONED flag, and there's a mediawiki/history/load job
that does that for all the other mediawiki tables.  For the tables used
in geowiki, this job does the same thing.  It first repairs the tables
to get them to recognize the new partitions, and then adds a
_PARTITIONED flag in the right place.

Bug: T190409
Change-Id: I7c41a77a69d0251a7ff255d92f603371616e0e58