Make mediawiki-history-reduced data permanent
This patch modifies the mediawiki-history-reduced oozie job so that
the generated data will be stored in a non-temporary hive table,
instead of a temporary one (deleted at the end of the job). The
ultimate goal is to be able to check data quality between multiple
snapshots.

This patch needs a companion in puppet to delete old snapshots.

Bug: T192482
Change-Id: Id31539df64b775bf984cecd0db136d3afe3f2d66