Correct webrequest dataloss false-positive script
The webrequest dataloss false-positive check was not incorrect
as the results it returned was valid, but there was a discrepancy
between the "wrong" numbers in webrequest_sequence_stats and
the ones in dataloss script.
The difference was due to webrequest_sequence_stats counting rows
with invalid timestamp (dt = '-') separately from the sequence
matching algorithm, while the dataloss script didn't.
It is interesting to understand the data-pattern of rows without dt in
hadoop partitions. Since varnishkafka assigns sequence nubmer to the
row at the beginning of the request and dt at the end of it, no-dt
rows in any given partition always can't have sequence numbers higher
than the max-sequence number for a with-dt row for that partition. Indeed
camus uses current-timestamp to partition rows that don't have a valid dt,
therefore those rows end up either in hour or hour+1, depending on when
it is generated and when camus runs.
This pattern explains why we filter no-dt rows in the current-hour but
not in the next-hour in the dataloss-script: for current hour, no-dt
rows have sequence numbers for either previous-jour or current-hour,
while next-hour has no-dt rows for current-hour and next-hour. The script
is actually interested in those rows with no-dt whith sequence numbers
belonging to current-hour.

Change-Id: I2b5ff3f5b1734f25397d9d70d636081a4eaae81c