Parameterize datasource of druid loading jobs
In our druid loading jobs, the target datasource was hard-coded
in the loading template.
This made testing jobs with non-prod datasources difficult.
Now the jobs have a oozie parameter overridable with
-Ddruid_datasource.

This patch also contains a new check in druid indexation
subworkflow so that non-hdfs users can only index datasources
with name starting with 'test_'.

Bug: T195882

Change-Id: I410c6118738732599ac0856e9497d484c2371c3f