Add validation step in mediawiki-history jobs
The denormalize oozie job now only cares the spark denormalization.
Hive partition addition happens after the new snapshot is validated
by the new check_denormalize job.
This new job uses spark to check user, page and denormalized
history, writing error-rows (if any) in a dedicated folder.
If the given folder is not empty after the job finishes,
errors have been found and the oozie job fails, not adding partitions
to the hive tables for the new snapshot. If no error folder is present,
new hive partitions are added to the tables and the other jobs flow.

Bug: T192481
Change-Id: I2ecc9c02f162b0195a64eafcad2bda390fa5327a