Update MWH-reduced to parquet storage
mediawiki_history_reduced was using json storage to be used as
druid indexation dataset. Since our last update, druid supports
parquet data indexation, so we update the base table.
There is a small gain in storage-size but the gain or other
analytics queries efficiency on the table will be substancial
(for instane the automtic data validation job).

Bug: T192483
Change-Id: Ibd30bd46a144f843c945d441f260232cdde9f804