Replace ids with text in mediawiki-history-reduced
We were using user_id and page_id to reference users and pages,
preventing easy-to-read user-facing results. This patch updates
the job to now index user_text (user_name or IP if anonymous) and
page_title (prefixed with canonical-namespace).
Values for aggregate have been checked and the differences are
unnoticeable (less than 10e-6 in most cases, 0.4% max), values
for top and count per-user or per-page have been checked as well
and are correct compared to wikistats1.
This path also involves druid index growth (we store strings
instead of integers): from ~250Gb to ~320Gb. Performance seems not
to have suffered on manual tests.

Bug: T201617
Change-Id: I2f44ae588d61ff92078701750c50a60de3801cda