Fix refinery-drop-druid-snapshots script
The --druid-host arg might be confusing since it seems
to ask a hostname:port combination but in reality it
requires a full URL. It also seems that a '/metadata/'
part of the URI is missing when trying to contact the
Druid coordinator for datasources. Last but not the least,
fix an exception raised with urlopen (json.loads cares about
str not bytes).

Change-Id: Ida9b0c7c711ad69229650e6c31b510dbf7afadcc