Set int namespace in hive mediawiki schemas
When sqooped, the data comes in as an integer, and expected so
by spark code (typed values). While reading bigint out of int
works in hive, if the schema is reused to generate dataset,
the typing breaks in spark. Making namespace int instead of
bigint seems the natural convention to follow.

Bug: T209179
Change-Id: I614f01fdd6de17b68fb7047a2aff9b2d00cce245