Update webrequest data-loss false positive check
After the problem of beginning of december with Varnish, we realized
the script checking for dataloss was incomplete and did not account
for some important cases: It worked weel in case some rows with no
timestamp (dt = '-') were present in the next hour, but did not work
weel when a lot more rows were shuffled in term of sequence numbers
(present in previous and next hours).
This patch updates the script to use spark2-sql instead of hive,
and check in a more advanced way if, by hosts, all rows defined
in an hour as the interval [MIN(sequence); MAX(sequence)] are
present when considereing previsous, current and next hour.
Note: Hosts with MIN(sequence) = 0 are discarded as it means their
varnish-kafka has restarted and sequence-flow is not longer valid
for the given hour.

Bug: T211000
Change-Id: I25900906de4e84b4cae6845bf455ba658765214d