Clickstream failing due to not being able to access hive tables
The clickstream job was configured to wait for data to be available
on HDFS, but not for the hive partitions to be created (while it
uses them). This patch changes the behavior for the job to wait on
the hive partitions to be created.

This patch also updates the jar version to be used by the job to
embed a fix on long->int for namespace, change:

https://gerrit.wikimedia.org/r/#/c/analytics/refinery/source/+/478983/

Bug: T211717
Change-Id: I4b1c302ebd413791ff1a1ff60c8045d945bcd5c3