Update mediawiki_history oozie job datasets
This patch adds a "unchecked" version of mediawiki_history
dataset in oozie, which is the state of the dataset before
the denormalize-check jobs have successfully run.
With this patch other jobs relying on mediawiki_history data
will only be started if the checker succeeds (it was already
the case for Hive partitions, but not for data).

Bug: T213524
Change-Id: I08e3955d9d285507a8038c59fb61a48f5f2d8e36