Use db_mapping to find the hostname
Refactor sqoop code to fetch the db cluster configuration and figure out
on what shard and port each db is on.  Also remove --jdbc-host parameter
from sqooping scripts.  This doesn't make sense as it only works in the
--labsdb case.  Instead, move towards having all db configuration
handled by util.py.  The caller only needs to pass whether they need the
labs or production replica cluster.  For now, I hardcoded the labsdb
host in util.py, but in the future this would ideally be a dns record
resolved in the same way as the production replica shareds are resolved.

Bug: T215290
Change-Id: Ibd9bfbfa43c677b7b5098f316ba5cceb87e0f01a