Fix daily and monthly partition thresholding
For daily or monthly partitions, the comparison:
 partition.datetime < threshold
was broken, because partition.datetime() returns the partition
start datetime, but what the script actually needs is the partition
end datetime. We didn't notice this problem before, because we were
not using daily or monthly granularities. The hourly partitions
were not having this problem, thanks to NOW being cropped to the
start of the hour.

This change implements thresholding correctly for hive partitions.
Plus adds some unit tests.

I have tested this new version against all data sets that are to
use this script and they all behaved as expected after the fix.

Bug: T226835
Change-Id: I5575f67166d6755e1b0adcd89517d4fbc5a5b05e