Remove duplicates from unique_devices_per_project_family
The grouping sets operation was working correctly on
unique_devices_per_domain, but it was creating duplicates
for unique_devices_per_project_family. The latter data
had already all records with access_method = 'all-sites'.
When grouped with grouping set (project, dt), it would
generate data with access_method = null, that later
would be coalesced to 'all-sites', which would be identical
to an already existing record.

This change does refactor the code a bit, to apply the
grouping by (project, dt) only to unique_devices_per_domain.

I did some tests and the results look good: identical
outputs except for the duplicates that are gone.

Bug: T229254
Change-Id: I6b193e8f566db6d3d9443e3fc2d86a1a37abce72