edit: set hive.exec.submit.local.task.via.child = false in the .hql file
With hive2 actions the hive.auto.convert.join setting causes some
extra work/commands to be processed by the hive-sever2's jvm. Any
command spawned will inherit the set of environments variables,
like -javaagent etc.. that creates the prometheus exporter (that
in turn tries to bind a specific port etc..). We disabled the
setting in hive-site.xml for T209536, but this coordinator needs
it since it is a huge performance booster.
The hive.exec.submit.local.task.via.child forces all the commands
to be executed in the HiveServer2's jvm, not triggering any failure.

Bug: T227257
Change-Id: Ifc4e15e4a432a69133fda03f2b3ee657c47532ca