Publish monthly geoeditor numbers
See discussion on task for background and privacy concerns and
mitigation.  Writes out a simpler, bucketed version of geoeditors
monthly to the hdfs archive.  From there, once pre-release tasks are
done, we can easily rsync the data to be served at dumps.wikimedia.org

Here, "bucketed" means adding a bit of imprecision to the numbers
published.  So if Romanian wikipedia has 5 editors from Hong Kong
with 100 or more edits, a row in the published file will be:
   rowiki	HK	100 or more	0	9

Deployment checklist:
* create geoeditors_blacklist_country table with provided script
* launch oozie coordinator, with a start date of 2018-01

Before this data is released we still have to:
* apply our new privacy and risk mitigation framework and publish results
* finalize the list of countries in the blacklist

Bug: T131280
Change-Id: I43e24d337ca9ba71f224282975f82e3c1ad5fbcb