Update geoeditors_daily - new dimensions and name
The geoeditors-daily table is used as an intermediate dataset.
As such, it should contain dimensions for bot identification
and action-type identification instead of filtering them out.
Also, the dataset is renamed editors_daily for the name to be
less restrictive than geoeditors as it contains more information
than just geolocation.

Bug: T238855
Change-Id: Id38ffdcdcabe3aa87ea3eb3b9c1eee63a86099e3