Table and workflow for features computations per actor per hour
We calculate every hour features that are of use to label
bot spikes (for bot traffic that is not self identified as such).

Note that actors here are defined in a simple way that makes
sense for labeling of bot spikes, it is not meant as a generic user or session
definition.

Bug: T238360
Change-Id: I4220e50823cdf67e09bab45e4ac4794a90744d75