Improve logging and return codes in refinery-drop-older-than
For some reason (I think it depends from running the unit tests
in the code) refinery-drop-older-than doesn't return 1 when an
exception is raised in main().

Moreover, when running the script without the checksum, it is nice
to get the list of directories to drop in separate lines, not in a single
one, to spot anomalies more easily (might be only me, in case I'll
drop the change).

Change-Id: Ia74edc23b4c75d79d1860b2d4f4ccf294d17c8df