refinery-drop-older-than: improve test for older_than arg
In the last change of refinery-drop-older-than the test was changed
to catch RuntimeError, more specific than Exception. This unveiled
a hidden behavior, namely that ValueError/TypeError exceptions
were caught too.

This change aims to use more specific exceptions in the code when
checking the validity of the older_than value.

Change-Id: Iae5072d0448f11feb1bb1769e96757a066327dda