oozie: Fix webrequest load workflow string
While testing Hue, I noticed that there is a problem in the
generation of the Workflow->Graph panel due to a json string encoding.
The XML of the workflow in fact is traslated into a JSON structure,
that surrounds everything with "", that leads to have:

{"subworkflow":{"app-path":"${replaceAll(wf:appPath(),"/[^/]*$","")}/check_sequence_statistics_workflow.xml"}}

The above is not valid JSON since the "" are not opened/closed/escaped
correctly. A quick fix to please Hue, and allow us to see the workflow's graph
when needed is using single quotes in the workflow.

I tested this patch in the Hadoop test cluster and I was able
to run the coordinator correctly. Hue was also correctly displaying
the Graph without emitting errors.

Change-Id: I0647b8d7d1c6be39d7cb31e994938103b38d6a36