Improve path discovery in drop-older-than
The deletion script always had a problem when the specified base_path
contains too many partitions. The way the paths are extracted to be
checked against the threshold involves using hdfs dfs -ls /*/*/.../*
When the script runs in prod, that is not a problem because year
folders are deleted before month folders (month folders before day
folders, etc.) and thus the glob targets are never too many. But when
the script runs in dry-run mode, no directories are deleted, and the
deeper the hdfs dfs -ls /*/*/.../* goes, the more costly the operation
is, to the point that the script fails. This is bad, because dry-run
mode is necessary to get the checksum for production execution.

Bug: T263495
Change-Id: Iccb637bf89dd63947e235c454d6ace95ac814bfa