Fix directory expansion bug in refinery-drop-older-than
To determine which sub-paths need to be deleted in an efficient way,
the script does not expand paths that are either too old that can
be deleted entirely, or too new that have to be kept entirely.
For instance: If the deletion threshold is 2020-03-15, the script
will mark a path like mydataset/year=2019 for deletion, without
expanding its monthly sub-paths, because all of that subtree must
be deleted. Similarly, a path like mydataset/year=2020/month=6
will be marked to be kept entirely, without expanding its daily
sub-paths. So, only directories that are "touching" the threshold
will be expanded.

Note that the script can only extract datetime information
from a path, once the path has been expanded to a certain depth.
For instance: Imagine the script is analyzing the path
mydataset/mycategory/hourly/year=2020/month=4/day=3/hour=12
It will start by executing hdfs.ls() on the first level mydataset/,
and will not be able to extract datetime information from that.
In this case, the script is marking the directory mydataset/ to be
expanded. On the next iteration, it will be expanded to
mydataset/mycategory, and so forth until the script reaches:
mydataset/mycategory/hourly/year=2020
only then, it will be able to extract datetime info from that,
and decide whether to delete it or not.

Now, the problem:
Depending on the regular expression that was given as an argument,
there can be some directories under base_path that do not match the
regular expression. For instance:
base_path=/wmf/data/event
path_format=(?!WMDEBanner)[A-Za-z]+/year=...
This regular expression excludes all datasets that have underscores
(and also the ones that start with WMDEBanner). We Analytics do not
want to delete those.
As the script is not able to extract datetime information from those
paths (because the regex doesn't ever match), it decides to further
expand those paths to look deeper for datetime information.
This makes it hdfs.ls() the whole tree, which can contain tens of
thousands of sub-paths.

The solution:
I added a method that tells whether a path is a partial match of the
regular expression. So that the script knows whether to continue
expanding a path, if no datetime info can be extracted from it yet.
This solves the problem, and does not introduce any risk, because
the addition only affects the paths that are to be expanded,
the paths that are to be deleted are assessed by another part
of the code. The new method is hacky indeed, but it solves the issue.
What do you think?

Bug: T263495
Change-Id: I1aafa4ed3808b3a341f0de32a1bb88d59848b39d